+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 
 - Fix release script of pushing release Tags
 - Fix plugin register related issue for all platforms
+- Fix register of all plugins with `enablePlugins: true` input
 
 ### Features